convert agentpoolprofile count to int64 to better correspond with underlying scaleset type

+++ pkg/api/admin/api/types.go
@@ -119,7 +119,7 @@ type RouterProfile struct {
 
 // MasterPoolProfile contains configuration for OpenShift master VMs.
 type MasterPoolProfile struct {
-	Count      *int    `json:"count,omitempty"`
+	Count      *int64  `json:"count,omitempty"`
 	VMSize     *VMSize `json:"vmSize,omitempty"`
 	SubnetCIDR *string `json:"subnetCidr,omitempty"`
 }
@@ -127,7 +127,7 @@ type MasterPoolProfile struct {
 // AgentPoolProfile represents configuration of OpenShift cluster VMs.
 type AgentPoolProfile struct {
 	Name       *string `json:"name,omitempty"`
-	Count      *int    `json:"count,omitempty"`
+	Count      *int64  `json:"count,omitempty"`
 	VMSize     *VMSize `json:"vmSize,omitempty"`
 	SubnetCIDR *string `json:"subnetCidr,omitempty"`
 	OSType     *OSType `json:"osType,omitempty"`
